refactor to app/types
